Cosmetics for #3079
