+++ celery/tests/bin/test_worker.py
@@ -4,8 +4,6 @@ import logging
 import os
 import sys
 
-from functools import wraps
-
 from billiard import current_process
 from kombu import Exchange, Queue
 
@@ -24,7 +22,7 @@ from celery.tests.case import (
     AppCase,
     Mock,
     SkipTest,
-    WhateverIO,
+    disable_stdouts,
     patch,
     skip_if_pypy,
     skip_if_jython,
@@ -38,25 +36,6 @@ class WorkerAppCase(AppCase):
     def tearDown(self):
         super(WorkerAppCase, self).tearDown()
         trace.reset_worker_optimizations()
-
-
-def disable_stdouts(fun):
-
-    @wraps(fun)
-    def disable(*args, **kwargs):
-        prev_out, prev_err = sys.stdout, sys.stderr
-        prev_rout, prev_rerr = sys.__stdout__, sys.__stderr__
-        sys.stdout = sys.__stdout__ = WhateverIO()
-        sys.stderr = sys.__stderr__ = WhateverIO()
-        try:
-            return fun(*args, **kwargs)
-        finally:
-            sys.stdout = prev_out
-            sys.stderr = prev_err
-            sys.__stdout__ = prev_rout
-            sys.__stderr__ = prev_rerr
-
-    return disable
 
 
 class Worker(cd.Worker):
@@ -210,20 +189,6 @@ class test_Worker(WorkerAppCase):
             self.assertTrue(worker.startup_info())
         finally:
             cd.ARTLINES = prev
-
-    @disable_stdouts
-    def test_startup_info_mongo_result_backend(self):
-        self.app.conf.result_backend = "mongodb://user:password@host0.com:43437,host1.com:43437/work4us?replicaSet=rs&ssl=true"
-        worker = self.Worker(app=self.app)
-        worker.on_start()
-        self.assertTrue(worker.startup_info())
-
-    @disable_stdouts
-    def test_startup_info_memcached_result_backend(self):
-        self.app.conf.result_backend = "cache+memcached://127.0.0.1:11211;127.0.0.2:11211;127.0.0.3/"
-        worker = self.Worker(app=self.app)
-        worker.on_start()
-        self.assertTrue(worker.startup_info())
 
     @disable_stdouts
     def test_run(self):
